fix(presetter): make runner resolution reliable
Previously, we use the command name as the runner name.
It allows us to change the cli name in package.json without
touching the code. However, it is unreliable.
e.g. The name runner name will be mistakenly resolved to `presetter`
if we invoke the runner by `npx presetter run <command>`.
Now, use the predetermined runner name to make it reliable.