fix(preset-web): auto find tailwind config file
NOTE: Without an explicit setting, the plugin would
only find the config file at the current working
directory, not the project root. This change would fix
the path finding mechanism such that the config can be
correctly located when it is sourced from other
locations such as a monorepo's root.